pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["egor <ekurlenko17@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "3.0.0b7"
apscheduler = "^3.10.1"
peewee = "^3.16.2"
aiomysql = "^0.2.0"
cryptography = "^41.0.7"
prettytable = "^3.9.0"
pylint = "^3.0.3"
mypy = "^1.8.0"
python-dotenv = "^1.0.1"
decouple = "^0.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
